requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
cryptography==42.0.7
longport==3.0.23
python-multipart==0.0.9
httpx==0.27.0
requests==2.31.0
boto3==1.34.100
numpy==1.26.4
pandas==2.2.1
ib_insync==0.9.86
nest-asyncio==1.6.0
psutil==5.9.8
yfinance==0.2.54
pytz==2024.1
duckdb==1.4.4
pyarrow==15.0.2

# Auth + PostgreSQL (web/Railway mode)
bcrypt==4.1.2
pyjwt==2.8.0
psycopg2-binary==2.9.9
